add pages and menu
