+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\React\\my-app\\src\\index.js":"1","D:\\React\\my-app\\src\\App.js":"2","D:\\React\\my-app\\src\\reportWebVitals.js":"3","D:\\React\\my-app\\src\\components\\HeaderComponent\\index.js":"4","D:\\React\\my-app\\src\\components\\FooterComponent\\index.js":"5","D:\\React\\my-app\\src\\components\\LayoutComponent\\index.js":"6"},{"size":219,"mtime":1611603420544,"results":"7","hashOfConfig":"8"},{"size":975,"mtime":1611780815118,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":499162500000,"results":"10","hashOfConfig":"8"},{"size":391,"mtime":1611778173718,"results":"11","hashOfConfig":"8"},{"size":311,"mtime":1611778315631,"results":"12","hashOfConfig":"8"},{"size":857,"mtime":1611780624250,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"ygcug0",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"21"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\React\\my-app\\src\\index.js",[],["28","29"],"D:\\React\\my-app\\src\\App.js",[],"D:\\React\\my-app\\src\\reportWebVitals.js",[],["30","31"],"D:\\React\\my-app\\src\\components\\HeaderComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\FooterComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\LayoutComponent\\index.js",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"32","replacedBy":"36"},{"ruleId":"34","replacedBy":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],["38"],["39"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\React\\my-app\\src\\index.js":"1","D:\\React\\my-app\\src\\App.js":"2","D:\\React\\my-app\\src\\reportWebVitals.js":"3","D:\\React\\my-app\\src\\components\\HeaderComponent\\index.js":"4","D:\\React\\my-app\\src\\components\\FooterComponent\\index.js":"5","D:\\React\\my-app\\src\\components\\LayoutComponent\\index.js":"6","D:\\React\\my-app\\src\\components\\PokemonCardComponent\\index.js":"7","D:\\React\\my-app\\src\\components\\MenuHeaderComponent\\index.js":"8","D:\\React\\my-app\\src\\components\\MenuComponent\\index.js":"9","D:\\React\\my-app\\src\\routes\\Home\\index.js":"10","D:\\React\\my-app\\src\\routes\\Game\\index.js":"11","D:\\React\\my-app\\src\\components\\NavBarComponent\\index.js":"12"},{"size":219,"mtime":1611603420544,"results":"13","hashOfConfig":"14"},{"size":551,"mtime":1612213373859,"results":"15","hashOfConfig":"14"},{"size":362,"mtime":499162500000,"results":"16","hashOfConfig":"14"},{"size":630,"mtime":1612210508283,"results":"17","hashOfConfig":"14"},{"size":311,"mtime":1611950418861,"results":"18","hashOfConfig":"14"},{"size":853,"mtime":1612205570037,"results":"19","hashOfConfig":"14"},{"size":2010,"mtime":1612205875306,"results":"20","hashOfConfig":"14"},{"size":454,"mtime":1612214366186,"results":"21","hashOfConfig":"14"},{"size":1069,"mtime":1612213709908,"results":"22","hashOfConfig":"14"},{"size":6347,"mtime":1612210636485,"results":"23","hashOfConfig":"14"},{"size":339,"mtime":1612213373853,"results":"24","hashOfConfig":"14"},{"size":737,"mtime":1612212828960,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"ygcug0",{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"33"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"50","messages":"51","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"D:\\React\\my-app\\src\\index.js",[],["52","53"],"D:\\React\\my-app\\src\\App.js",[],"D:\\React\\my-app\\src\\reportWebVitals.js",[],["54","55"],"D:\\React\\my-app\\src\\components\\HeaderComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\FooterComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\LayoutComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\PokemonCardComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\MenuHeaderComponent\\index.js",[],"D:\\React\\my-app\\src\\components\\MenuComponent\\index.js",[],"D:\\React\\my-app\\src\\routes\\Home\\index.js",[],"D:\\React\\my-app\\src\\routes\\Game\\index.js",[],"D:\\React\\my-app\\src\\components\\NavBarComponent\\index.js",["56"],{"ruleId":"57","replacedBy":"58"},{"ruleId":"59","replacedBy":"60"},{"ruleId":"57","replacedBy":"61"},{"ruleId":"59","replacedBy":"62"},{"ruleId":"63","severity":1,"message":"64","line":18,"column":17,"nodeType":"65","endLine":19,"endColumn":75},"no-native-reassign",["66"],"no-negated-in-lhs",["67"],["66"],["67"],"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
