+++ src/components/MenuHeaderComponent/style.module.css
@@ -1,0 +1,90 @@
+.root {
+    background: #202736;
+    background-attachment: fixed;
+    background-size: cover;
+    position: relative;
+    min-height: 100vh;
+    margin: 0 auto;
+    z-index: 1;
+}
+
+.root::before {
+    content: '';
+    position: fixed;
+    background: url('../../assets/Stars.svg');
+    background-attachment: fixed;
+    width: 100%;
+    min-height: 100vh;
+    z-index: -1;
+    animation: stars-move-in 1000ms 300ms forwards;
+    opacity: 0;
+}
+
+@keyframes stars-move-in {
+    from {
+        background-position-y: -100px;
+    }
+    to {
+        opacity: 1;
+        background-position-y: 0;
+    }
+}
+
+.forest {
+    position: absolute;
+    bottom: -300px;
+    left: 0;
+    background: url('../../assets/Trees.svg') bottom left repeat-x;
+    background-size: cover;
+    width: 100%;
+    height: 80%;
+    opacity: 0;
+    animation: forest-move-in 1000ms 500ms forwards;
+    border-bottom: 300px solid #181d23;
+}
+
+@keyframes forest-move-in {
+    from {
+        background-position-y: 150%;
+    }
+    to {
+        opacity: 1;
+        background-position-y: 100%;
+    }
+}
+
+.container {
+    width: fit-content;
+    position: absolute;
+    right: 0;
+    top: 50%;
+    left: 50%;
+    opacity: 0;
+    transform: translate(-50%, -50%);
+    animation: text-fade-in 1000ms 800ms forwards;
+}
+
+@keyframes text-fade-in {
+    from {
+        right: 0;
+    }
+    to {
+        opacity: 1;
+        right: 25%;
+    }
+}
+
+.container h1 {
+    font-size: 65px;
+    font-weight: normal;
+    font-style: italic;
+    color: #fafafa;
+    line-height: 65px;
+}
+
+.container p {
+    font-size: 28px;
+    font-weight: normal;
+    color: #fafafa;
+    line-height: 32px;
+}
